+++ src/com/allen_sauer/gwt/game/client/ui/util/MathUtil.java
@@ -4,7 +4,7 @@
 package com.allen_sauer.gwt.game.client.ui.util;
 
 public final class MathUtil {
-  public static final int sign(int val) {
+  public static final int sign(double val) {
     if (val > 0) {
       return 1;
     } else if (val < 0) {
